requirements.txt
Django==4.0.5
django-debug-toolbar==3.4.0
django-phonenumber-field==6.3.0
environs==9.5.0
geopy==2.2.0
Jinja2==3.1.2
lxml==4.9.0
phonenumbers==8.12.50
phonenumberslite==8.12.51
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.1
qrcode==7.3.1